backend/app/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
requests
tensorflow-cpu
tensorflow-hub
scikit-learn
python-multipart
python-dotenv